feat: Possibly fixed some of the failing tests in CI pipeline (#7)
* feat: Possibly fixed some of the failing tests in CI pipeline

- It seems the reason why a good chunk of the tests that run on
the CI pipeline fail is because of the version of Ubuntu (20.0.4, focal)
that the tests run within. Ubuntu focal uses curl v7.68.x which does not include
the `--output-dir` option included in curl v7.74.x.
This is actually a compatibility issue since not many people may have the most
up-to-date curl version. There are a couple of ways to fix this but the
optimal solution would be to have the `installWithCurl` fallback to
`installWithWget` on failure since `wget`'s capabilities
(as leveraged in the source code) is more mature than curl

- Added fallback to downloading with wget if curl download fails

* chore: Added retry cap for wget download during tests

* fix: Fixed zx learning tests and made sure tests run in correct dir

- Reduced cap on wget retries to 1 to prevent slow tests

* fix: Broken import in app.test.ts

* fix: reset working dir on EACH test run

* ci: Added npm script to run tests in CI environment

- The difference between the CI script and the normal test script
is the verbosity level. The CI script is more verbose

* ci: Replaced husky with github actions